+++ apps/blog-cms/src/containers/Post/components/PostTable.tsx
@@ -281,7 +281,11 @@ const PostTable: FC<Props> = ({
 
         {total === 0 || (
           <div className={classes.pagination}>
-            <Select value={pageSize.toString()} onChange={handlePageSizeChange}>
+            <Select
+              value={pageSize.toString()}
+              onChange={handlePageSizeChange}
+              className={classes.selector}
+            >
               <MenuItem value={10}>10</MenuItem>
               <MenuItem value={20}>20</MenuItem>
               <MenuItem value={50}>50</MenuItem>