+++ apps/blog-web/src/containers/Home/components/PostList.tsx
@@ -1,75 +1,100 @@
 import { FC, useState } from 'react'
-import { useLazyQuery } from '@apollo/client'
+import { print, ASTNode } from 'graphql'
+// import { useLazyQuery } from '@apollo/client'
 import { SVG_SPRITE } from 'src/shared/constants'
 import PostCard from 'src/containers/Post/components/PostCard/PostCard'
-import PostCardSkeleton from 'src/containers/Post/components/PostCardSkeleton/PostCardSkeleton'
-import SkeletonIterator from 'src/components/SkeletonIterator/SkeletonIterator'
 import { POSTS } from 'src/containers/Post/typeDefs'
-import { PostQuery, PostVars, IPostItem } from 'src/containers/Post/types'
+import {
+  // PostQuery,
+  // PostVars,
+  IPostItem,
+  IPost
+} from 'src/containers/Post/types'
 import InfiniteScroll from 'src/components/InfiniteScroll/InfiniteScroll'
 import SubTitle from './SubTitle'
 import { Status } from '../styled'
 
-interface List {
-  posts: IPostItem[]
+interface Props {
+  data: IPost
 }
 
-const PostList: FC = () => {
-  const [data, setData] = useState<List>({ posts: [] })
-  const [page, setPage] = useState(1)
+const PostList: FC<Props> = ({ data: ssrData }) => {
+  const [data, setData] = useState<IPostItem[]>(ssrData.items)
+  const [page, setPage] = useState(2)
   const [hasMore, setHasMore] = useState(true)
+  const [loading, setLoading] = useState(false)
 
-  const [getPosts, { loading }] = useLazyQuery<PostQuery, PostVars>(POSTS, {
-    onCompleted(continuousPosts) {
-      const { items, total, page: currentPage } = continuousPosts.posts
-      if (Math.ceil(total / 10) === currentPage) {
-        setHasMore(false)
-      }
-      setPage((_page) => _page + 1)
-      setData({ posts: data.posts.concat(items) })
-    }
-  })
+  // const [fetchMore, { loading }] = useLazyQuery<PostQuery, PostVars>(POSTS, {
+  //   onCompleted(continuousPosts) {
+  //     const { items, total, page: currentPage } = continuousPosts.posts
+  //     setData([...data, ...items])
+  //     if (Math.ceil(total / 10) === currentPage) {
+  //       setHasMore(false)
+  //       return
+  //     }
+  //     setPage(currentPage + 1)
+  //   }
+  // })
 
-  const fetchPosts = (tag?: string) => {
-    if (loading) return
+  // const fetchPosts = (tag?: string) => {
+  //   fetchMore({
+  //     variables: {
+  //       input: {
+  //         page,
+  //         pageSize: 10,
+  //         tag
+  //       }
+  //     }
+  //   })
+  // }
 
-    getPosts({
-      variables: {
-        input: {
-          page,
-          pageSize: 10,
-          tag
+  const fetchData = () => {
+    setLoading(true)
+    fetch(process.env.NEXT_PUBLIC_API_URL, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        query: print(POSTS as ASTNode),
+        variables: {
+          input: {
+            page,
+            pageSize: 10
+          }
         }
-      }
+      })
     })
+      .then((res) => res.json())
+      .then((result) => {
+        const { items, total, page: currentPage } = result.data.posts as IPost
+        setData([...data, ...items])
+        if (Math.ceil(total / 10) === currentPage) {
+          setHasMore(false)
+          return
+        }
+        setPage(currentPage + 1)
+      })
+      .finally(() => {
+        setLoading(false)
+      })
   }
 
   return (
     <InfiniteScroll
       hasMoreData={hasMore}
       isLoading={loading}
-      onBottomHit={fetchPosts}
-      loadOnMount
+      onBottomHit={fetchData}
+      loadOnMount={false}
     >
       <SubTitle icon={SVG_SPRITE.new} title="The Latest!" />
 
-      {data.posts.length === 0 ? (
-        <SkeletonIterator count={3} skeletonComponent={PostCardSkeleton} />
-      ) : (
-        data.posts.map((post: IPostItem) => (
-          <PostCard post={post} key={post._id} />
-        ))
-      )}
+      {data.map((post: IPostItem) => (
+        <PostCard post={post} key={post._id} />
+      ))}
 
       <Status>
-        {
-          // eslint-disable-next-line no-nested-ternary
-          data.posts.length > 0 && loading
-            ? '正在加载中...'
-            : hasMore
-            ? ''
-            : '没有更多了...'
-        }
+        {data.length > 0 && hasMore ? '正在加载中...' : '没有更多了...'}
       </Status>
     </InfiniteScroll>
   )