+++ apps/blog-web/src/containers/Layout/Layout.tsx
@@ -1,9 +1,8 @@
-import { FC, useState, useEffect } from 'react'
+import { FC, useEffect } from 'react'
 import { useQuery } from '@apollo/client'
-import { hotjar } from 'react-hotjar'
-import throttle from 'lodash.throttle'
+import NProgress from 'nprogress'
+import { useRouter } from 'next/router'
 import { initGA, logPageView } from 'src/shared/analytics'
-import { BACK_TO_TOP_THRESHOLD } from 'src/shared/constants'
 import SVGSprite from 'src/components/SVGSprite/SVGSprite'
 import AlgoliaSearchBox from 'src/containers/Post/components/AlgoliaSearchBox/AlgoliaSearchBox'
 import { GET_GLOBAL_SETTING } from 'src/containers/GlobalSetting/typeDefs'
@@ -26,40 +25,27 @@ interface Props {
 }
 
 const Layout: FC<Props> = ({ title, children }) => {
+  const router = useRouter()
   const { data } = useQuery<GlobalSettingQuery>(GET_GLOBAL_SETTING)
 
-  const [scrollTopCount, setScrollTopCount] = useState(0)
-
-  const scrollTopCountHandler = throttle(() => {
-    const top = document.documentElement.scrollTop || document.body.scrollTop
-    setScrollTopCount(top)
-  }, 100)
-
   useEffect(() => {
-    document.addEventListener('scroll', scrollTopCountHandler, {
-      passive: true
-    })
-
-    return () => {
-      document.removeEventListener('scroll', scrollTopCountHandler)
-    }
-  }, [])
-
-  useEffect(() => {
-    if (process.env.NODE_ENV === 'production') {
+    if (process.env.NODE_ENV === 'production' && window !== undefined) {
       if (!window.GA_INITIALIZED) {
         initGA()
         window.GA_INITIALIZED = true
+      } else {
+        logPageView()
       }
-
-      logPageView()
-
-      hotjar.initialize(
-        parseInt(process.env.NEXT_PUBLIC_HOTJAR_ID, 10),
-        parseInt(process.env.NEXT_PUBLIC_HOTJAR_SV, 10)
-      )
     }
   }, [])
+
+  useEffect(() => {
+    router.events.on('routeChangeStart', () => {
+      NProgress.start()
+    })
+    router.events.on('routeChangeComplete', () => NProgress.done())
+    router.events.on('routeChangeError', () => NProgress.done())
+  }, [router.events])
 
   return (
     <Layouts>
@@ -73,7 +59,7 @@ const Layout: FC<Props> = ({ title, children }) => {
       />
       <Nav />
       <SVGSprite />
-      <BackToTop isShowCat={scrollTopCount >= BACK_TO_TOP_THRESHOLD} />
+      <BackToTop />
       <AlgoliaSearchBox />
     </Layouts>
   )