+++ apps/blog-web/src/shared/constants.ts
@@ -97,3 +97,10 @@ export const months = [
   'Nov',
   'Dec'
 ]
+
+export const APLAYER_CDN =
+  'https://cdn.jsdelivr.net/npm/aplayer@1.10.1/dist/APlayer.min.js'
+
+  export const DOMAIN = 'https://yanceyleo.com'
+
+  export const RSS = `${DOMAIN}/rss.xml`