+++ apps/uploader-service/src/libs/guard/auth.guard.ts
@@ -6,9 +6,9 @@ export const AuthModule = KeycloakConnectModule.registerAsync({
     const KEY_CLOAK_URL = configService.get<string>('KEY_CLOAK_URL')
     const KEY_CLOAK_REALM = configService.get<string>('KEY_CLOAK_REALM')
     const KEY_CLOAK_CLIENT_ID = configService.get<string>(
-      'KEY_CLOAK_CLIENT_SECRET'
+      'KEY_CLOAK_CLIENT_ID'
     )
-    const KEY_CLOAK_CLIENT_SECRET = configService.get<string>('KEY_CLOAK_URL')
+    const KEY_CLOAK_CLIENT_SECRET = configService.get<string>('KEY_CLOAK_CLIENT_SECRET')
     return {
       authServerUrl: KEY_CLOAK_URL,
       realm: KEY_CLOAK_REALM,