+++ apps/uploader-service/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - uploader-service
 
-This log was last generated on Mon, 29 Nov 2021 04:57:53 GMT and should not be manually modified.
+This log was last generated on Sun, 09 Jan 2022 14:46:08 GMT and should not be manually modified.
+
+## 1.1.1
+Sun, 09 Jan 2022 14:46:08 GMT
+
+### Patches
+
+- fix env params error
 
 ## 1.1.0
 Mon, 29 Nov 2021 04:57:53 GMT